Move to resource utils
