+++ src/poly/tests/project_test.py
@@ -32,10 +32,10 @@ from poly.resources import (
     SettingsRole,
     SettingsRules,
     SMSTemplate,
-    Topic,
     TestCase,
     TestCaseAssertion,
     TestCaseTags,
+    Topic,
     TranscriptCorrection,
     Translation,
     Variable,
